MDLSITE-4419: Move all scripts to common prepare_npm_stuff.sh
New behavior is simple:

- if npmbase is passed, then the script will call to
  prepare_npm_stuff.sh and also clean link.
- if npmbase is not passed, then it will assume somebody else
  (caller) has prepared the npm stuff already and won't
  operate anything.

+++ grunt_process/grunt_process.sh
@@ -4,7 +4,7 @@
 # $gitbranch: Branch we are going to install the DB
 # $extrapath: Extra paths to be available (global)
 # $npmcmd: Path to the npm executable (global)
-# $npmbase: Base directory where we'll store multiple npm packages versions (subdirectories per branch)
+# $npmbase: (optional) Base directory where we'll store multiple npm packages versions (subdirectories per branch)
 
 # Let's be strict. Any problem leads to failure.
 set -e
@@ -14,7 +14,7 @@ if [[ -n ${extrapath} ]]; then
     export PATH=${PATH}:${extrapath}
 fi
 
-required="WORKSPACE gitdir gitbranch extrapath npmcmd npmbase"
+required="WORKSPACE gitdir gitbranch extrapath npmcmd"
 for var in $required; do
     if [ -z "${!var}" ]; then
         echo "Error: ${var} environment variable is not defined. See the script comments."
@@ -34,35 +34,11 @@ rm -fr config.php
 rm -fr ${outputfile}
 rm -fr ${outputfile}.stderr
 
-# Verify we have the npmbase dir, creating if needed
-if [[ ! -d ${npmbase} ]]; then
-    echo "WARN: npmbase dir (${npmbase}) not found. Creating it"
-    mkdir -p ${npmbase}
-    echo "NOTE: npmbase dir (${npmbase}) created"
-else
-    echo "OK: npmbase dir (${npmbase}) found"
+# Prepare all the npm stuff if needed
+# (only if the job is in charge of handling it, aka, $npmbase was passed
+if [[ -n ${npmbase} ]]; then
+    ${mydir}/../prepare_npm_stuff/prepare_npm_stuff.sh
 fi
-
-# Verify we have already the gitbranch dir, creating if needed
-if [[ ! -d ${npmbase}/${gitbranch}/node_modules ]]; then
-    echo "WARN: npmbase for branch (${gitbranch}) not found. Creating it"
-    mkdir -p ${npmbase}/${gitbranch}/node_modules
-    echo "NOTE: npmbase for branch (${gitbranch}) created"
-else
-    echo "OK: npmbase for branch (${gitbranch}) found"
-fi
-
-# Linking it.
-ln -nfs ${npmbase}/${gitbranch}/node_modules ${gitdir}/node_modules
-
-# Verify there is a grunt executable available, installing if missing
-if [[ ! -f ${gitdir}/node_modules/grunt-cli/bin/grunt ]]; then
-    echo "WARN: grunt-cli executable not found. Installing everything"
-    ${npmcmd} install grunt-cli
-fi
-
-# Always run npm install to keep our npm packages correct
-${npmcmd} install
 
 # Run grunt against the git repo
 cd ${gitdir}
@@ -72,11 +48,22 @@ rm -fr $(find . -path '*/amd/build' -type d)
 
 # Send both stdout and stderr to files while passing them intact (for callers consumption).
 # The echo here works around a problem where shifter is sending colours (MDL-52591).
-echo | ${gitdir}/node_modules/grunt-cli/bin/grunt --no-color > >(tee "${outputfile}") 2> >(tee "${outputfile}".stderr >&2)
-exitstatus=${PIPESTATUS[0]}
+gruntcmd="$(${npmcmd} bin)"/grunt
+if [ -x $gruntcmd ]; then
+    set +e
+    $gruntcmd --no-color > >(tee "${outputfile}") 2> >(tee "${outputfile}".stderr >&2)
+    set -e
+    exitstatus=$?
+else
+    echo "Error: grunt executable not found" | tee "${outputfile}"
+    exit 1
+fi
 
 # Cleanup symlink as not required after run (and prevent other jobs operating on it)
-rm ${gitdir}/node_modules
+# (only if the job is in charge of handling it, aka, $npmbase was passed
+if [[ -n ${npmbase} ]]; then
+    rm ${gitdir}/node_modules
+fi
 
 if [ $exitstatus -ne 0 ]; then
     echo "ERROR: Problems running grunt" | tee -a "${outputfile}"
