+++ shifter_walk/shifter_walk.sh
@@ -1,13 +1,24 @@
 #!/bin/bash
+# $WORKSPACE: Directory where results/artifacts will be created
 # $gitdir: Directory containing git repo
 # $gitbranch: Branch we are going to install the DB
 # $extrapath: Extra paths to be available (global)
+# $gitcmd: Path to the git executable (global)
 # $npmcmd: Path to the npm executable (global)
 # $shifterbase: Base directory where we'll store multiple shifter versions (can be different by branch)
-# $shifterversion: Version of shifter to be used by this job
+# $shifterversion: (optional) Version of shifter to be used by this job
 
 # Let's be strict. Any problem leads to failure.
 set +e
+
+# Verify everything is set
+required="WORKSPACE gitdir gitbranch extrapath gitcmd npmcmd shifterbase"
+for var in ${required}; do
+    if [ -z "${!var}" ]; then
+        echo "Error: ${var} environment variable is not defined. See the script comments."
+        exit 1
+    fi
+done
 
 # Let add $path to PATH
 if [[ -n ${extrapath} ]]; then
@@ -25,41 +36,27 @@ cd ${gitdir} && git reset --hard ${gitbranch}
 rm -fr config.php
 rm -fr ${outputfile}
 
-# Verify we have the shifterbase dir, creating if needed
-if [[ ! -d ${shifterbase} ]]; then
-    echo "WARN: shifterbase dir (${shifterbase}) not found. Creating it"
-    mkdir -p ${shifterbase}
-    echo "NOTE: shifterbase dir (${shifterbase}) created"
-else
-    echo "OK: shifterbase dir (${shifterbase}) found"
-fi
-
-# Verify we have already the shifterversion dir, creating if needed
-if [[ ! -d ${shifterbase}/${shifterversion} ]]; then
-    echo "WARN: shifterversion dir (${shifterversion}) not found. Creating it"
-    mkdir -p ${shifterbase}/${shifterversion}
-    echo "NOTE: shifterversion dir (${shifterversion}) created"
-else
-    echo "OK: shifterversion dir (${shifterversion}) found"
-fi
-
-# Verify there is a shifter executable available, installing id neeed
-if [[ ! -f ${shifterbase}/${shifterversion}/node_modules/shifter/bin/shifter ]]; then
-    echo "WARN: shifter executable (${shifterversion}) not found. Installing it"
-    cd ${shifterbase}/${shifterversion}
-    ${npmcmd} install shifter@${shifterversion}
-    echo "NOTE: shifter executable (${shifterversion}) installed"
-else
-    echo "OK: shifter executable (${shifterversion}) found"
-fi
+# Prepare all the npm stuff if needed
+export npmbase=${shifterbase}
+${mydir}/../prepare_npm_stuff/prepare_npm_stuff.sh
 
 # Run shifter against the git repo
 cd ${gitdir}
+
+shiftercmd="$(cd ${shifterbase}/${gitbranch}/node_modules && ${npmcmd} bin)"/shifter
+if [ ! -x $shiftercmd ]; then
+    echo "Error: shifter executable not found" | tee "${outputfile}"
+    exit 1
+fi
+
 # First delete all shifted files so we can detect stale files in the build dir
 rm -fr `find . -path '*/yui/build' -type d`
 
-${shifterbase}/${shifterversion}/node_modules/shifter/bin/shifter --walk --recursive | tee "${outputfile}"
+set +e
+${shiftercmd} --walk --recursive | tee "${outputfile}"
 exitstatus=${PIPESTATUS[0]}
+set -e
+
 if [ $exitstatus -ne 0 ]; then
     echo "ERROR: Problems running shifter" | tee -a "${outputfile}"
     exit $exitstatus
